jda/requirements.txt
Django==1.9
django-cors-headers==1.1.0
djangorestframework==3.3.3
Markdown==2.6.6
MySQL-python==1.2.5
